refactor(#66): 버튼 클릭 범위 수정 및 BittonBox component 위치 변경
ButtonBox component 위치 변경
